requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
google-generativeai==0.5.2
Werkzeug==3.0.1
gunicorn==21.2.0
markdown==3.6
pdfkit==1.0.0
python-dotenv==1.0.1
redis==5.0.1
email-validator==2.1.0
cryptography>=41.0.0
WTForms==3.1.1
beautifulsoup4>=4.14.0
html2text>=2025.4.15
reportlab>=4.4.4
whitenoise>=6.5.0
psycopg2-binary>=2.9.9
pillow>=10.0.0
